Rename CRD group name
