+++ operator/config/webhook/webhook.yaml
@@ -1,7 +1,7 @@
 apiVersion: admissionregistration.k8s.io/v1
 kind: MutatingWebhookConfiguration
 metadata:
-  name: defaulting.webhook.controlplane.kit.sh
+  name: defaulting.webhook.controlplane.kit.k8s.sh
 webhooks:
 - admissionReviewVersions: ["v1"]
   clientConfig:
@@ -10,10 +10,10 @@ webhooks:
       namespace: kit
   failurePolicy: Fail
   sideEffects: None
-  name: defaulting.webhook.controlplane.kit.sh
+  name: defaulting.webhook.controlplane.kit.k8s.sh
   rules:
   - apiGroups:
-    - kit.k8s.amazonaws.com
+    - kit.k8s.sh
     apiVersions:
     - v1alpha1
     resources:
@@ -28,7 +28,7 @@ webhooks:
 apiVersion: admissionregistration.k8s.io/v1
 kind: ValidatingWebhookConfiguration
 metadata:
-  name: validation.webhook.controlplane.kit.sh
+  name: validation.webhook.controlplane.kit.k8s.sh
 webhooks:
 - admissionReviewVersions: ["v1"]
   clientConfig:
@@ -37,10 +37,10 @@ webhooks:
       namespace: kit
   failurePolicy: Fail
   sideEffects: None
-  name: validation.webhook.controlplane.kit.sh
+  name: validation.webhook.controlplane.kit.k8s.sh
   rules:
   - apiGroups:
-    - kit.k8s.amazonaws.com
+    - kit.k8s.sh
     apiVersions:
     - v1alpha1
     resources:
