+++ operator/docs/examples/cluster.yaml
@@ -1,4 +1,4 @@
-apiVersion: kit.k8s.amazonaws.com/v1alpha1
+apiVersion: kit.k8s.sh/v1alpha1
 kind: ControlPlane
 metadata:
   name: example # Desired Cluster Name
