+++ operator/pkg/apis/controlplane/v1alpha1/controlplane_defaults.go
