+++ operator/pkg/apis/controlplane/v1alpha1/doc.go
@@ -16,7 +16,7 @@ limitations under the License.
 // +k8s:openapi-gen=true
 // +k8s:deepcopy-gen=package,register
 // +k8s:defaulter-gen=TypeMeta
-// +groupName=kit.k8s.amazonaws.com
+// +groupName=kit.k8s.sh
 package v1alpha1
 
 import (
@@ -32,7 +32,7 @@ var (
 
 	ControlPlaneKind = "ControlPlane"
 	// SchemeGroupVersion is group version used to register these objects
-	SchemeGroupVersion = schema.GroupVersion{Group: "kit.k8s.amazonaws.com", Version: APIVersion}
+	SchemeGroupVersion = schema.GroupVersion{Group: "kit.k8s.sh", Version: APIVersion}
 
 	// SchemeBuilder is used to add go types to the GroupVersionKind scheme
 	SchemeBuilder = &scheme.Builder{GroupVersion: SchemeGroupVersion}
