+++ operator/pkg/utils/reconciler/reconciler.go
@@ -17,7 +17,7 @@ package reconciler
 import (
 	"context"
 
-	"github.com/awslabs/kit/operator/pkg/apis/infrastructure/v1alpha1"
+	"github.com/awslabs/kit/operator/pkg/apis/controlplane/v1alpha1"
 )
 
 type Interface interface {
